service/pyproject.toml
[tool.poetry]
name = "pixelizer-service"
package-mode = false
version = "0.1.0"
description = ""
authors = ["nifle3 <kuper116rus.2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.4"
loguru = "^0.7.2"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
